% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abvd.feature.R
\name{abvd.feature}
\alias{abvd.feature}
\title{Download ABVD data}
\usage{
abvd.feature(feature, glottolog.source = "modified")
}
\arguments{
\item{feature}{A character vector that define a language id from ABVD (e. g. "1", "292").}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
This function downloads data from ABVD (https://abvd.shh.mpg.de/austronesian/) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# abvd.feature(c(292, 7))
}
\seealso{
\code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{wals.feature}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
