% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abvd.R
\docType{data}
\name{abvd}
\alias{abvd}
\title{ABVD's Language identifiers}
\format{
A data frame with 1468 rows and 2 variables:
\describe{
  \item{id}{language identifier}
  \item{glottocode}{Glottocode}
}
}
\usage{
abvd
}
\description{
Language identifiers from ABVD (\url{https://abvd.eva.mpg.de/austronesian/}). This dataset is created for \code{\link{abvd.feature}} function.
}
\keyword{datasets}
