% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanuatu.feature.R
\name{add_param}
\alias{add_param}
\title{Download Vanuatu Voices data}
\usage{
add_param(x)
}
\arguments{
\item{feature}{A string with a parameter from Conepts (\url{https://vanuatuvoices.clld.org/parameters}))}

\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}
}
\description{
This function downloads data from Vanuatu Voices (\url{https://vanuatuvoices.clld.org/}). You need the internet connection.
}
\author{
Mikhail Leonov
}
