% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afbo.feature.R
\name{afbo.feature}
\alias{afbo.feature}
\title{Download AfBo data}
\usage{
afbo.feature(features = "all", glottolog.source = "modified")
}
\arguments{
\item{features}{A character vector that define with an affix functions from AfBo (e. g. "all", "adjectivizer", "focus").}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
This function downloads data from AfBo (http://afbo.info) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# afbo.feature()
# afbo.feature(c("adjectivizer", "focus"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
