% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.lang.R
\name{aff.lang}
\alias{aff.lang}
\title{Get affiliation by languoid}
\usage{
aff.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return affiliation.
}
\author{
George Moroz <agricolamz@gmail.com>
}

