% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotyp.R
\docType{data}
\name{autotyp}
\alias{autotyp}
\title{AUTOTYP's Language identifiers}
\format{
An object of class \code{data.frame} with 2950 rows and 2 columns.
}
\usage{
autotyp
}
\description{
Language identifiers from AUTOTYP v. 0.1.0 (\url{https://github.com/autotyp/autotyp-data}). This dataset is created for \code{\link{autotyp.feature}} function.
}
\details{
#' @format A data frame with 2950 rows and 2 variables:
\describe{
  \item{LID}{language identifier}
  \item{Glottocode}{Glottocode}
}
}
\keyword{datasets}
