% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantu.feature.R
\name{bantu.feature}
\alias{bantu.feature}
\title{Download BANTU data}
\usage{
bantu.feature(features)
}
\arguments{
\item{features}{A character vector that define with a feature ids from BANTU ('house', 'cat').}
}
\description{
This function downloads data from Bantu Basic Vocabulary Database (\url{https://abvd.eva.mpg.de/bantu/index.php}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# bantu.feature(c('house', 'cat'))
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}
}
\author{
Anna Smirnova <annedadaa@gmail.com>
}
