% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circassian.R
\docType{data}
\name{circassian}
\alias{circassian}
\title{Circassian villages in Russia}
\format{A data frame with 8286 rows and 7 variables:
\describe{
  \item{longitude}{longitude}
  \item{latitude}{latitude}
  \item{village}{name of the village}
  \item{district}{names of the subjects of the Russian Federation: kbr --- Kabardino-Balkar Republic, kch --- Karachay-Cherkess Republic, kk --- Krasnodar Krai, ra --- Republic of Adygea, stv --- Stavropol Krai}
  \item{languoid}{names of the Circassian dialects}
  \item{language}{according standard Circassian devision there are Adyghe and Kabardian languages}
}}
\usage{
circassian
}
\description{
A dataset containes the list of the Circassian villages in Russia
with genealogical affiliation, coordinates and district names. Most
data collected during the fieldworks (2011-2016).
}
\keyword{datasets}

