% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Catalogue of countries}
\format{
A data frame with 189 rows and 5 variables:
\describe{
  \item{alpha3}{ISO 3166-3 code of the country}
  \item{alpha2}{ISO 3166-2 code of the country}
  \item{country_name}{Country name}
  \item{additional_names}{Additional names of the country}
  \item{official_languages}{Official languages}
}
}
\usage{
countries
}
\description{
Catalogue of countries, ISO-codes and official languages
}
\keyword{datasets}
