% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cow.R
\name{cow}
\alias{cow}
\title{Sounds of a cow
This function produces a specified number of Sounds mu}
\usage{
cow(num)
}
\arguments{
\item{num}{times of repetitions}
}
\description{
Sounds of a cow
This function produces a specified number of Sounds mu
}
\examples{
# cow()
}
\author{
Sasha Shakhnova
}
