% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejective_and_n_consonants.R
\docType{data}
\name{ejective_and_n_consonants}
\alias{ejective_and_n_consonants}
\title{Number of consonants and presence of ejectives}
\format{A data frame with 19 rows and 4 variables:
\describe{
  \item{language}{language name}
  \item{consonants}{number of consonants. Based on UPSID database.}
  \item{vowels}{number of vowels. Based on UPSID database.}
  \item{ejectives}{presence of ejective sounds}
}}
\usage{
ejective_and_n_consonants
}
\description{
Number of consonants and presence of ejectives
}
\keyword{datasets}
