% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurasianphonology.feature.R
\name{eurasianphonolgy.feature}
\alias{eurasianphonolgy.feature}
\title{Open .csv file of The database of Eurasian phonological inventories (beta version) which was downloaded via eurasianphonology.R script.}
\usage{
eurasianphonolgy.feature()
}
\description{
This function open downloaded data from the database of Eurasian phonological inventories (\url{http://eurasianphonology.info}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
eurasianphonolgy.feature

}
\seealso{
\code, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
