% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurasianphonology.feature.R
\name{eurasianphonology.feature}
\alias{eurasianphonology.feature}
\title{Download data from the database of Eurasian phonological inventories}
\usage{
eurasianphonology.feature()
}
\description{
This function open downloaded data from the database of Eurasian phonological inventories (\url{http://eurasianphonology.info}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{

eurasianphonology.feature()

}
\seealso{
\code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
