% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurasianphonology.R
\docType{data}
\name{eurasianphonology}
\alias{eurasianphonology}
\title{Eurasianphonology data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19825 rows and 19 columns.
}
\usage{
eurasianphonology
}
\description{
Data from The database of Eurasian phonological inventories (\url{http://eurasianphonology.info}). This dataset is created for \code{\link{eurasianphonology.feature}} function.
}
\details{
#' @format A data frame with 19825 rows and 19 variables:
\describe{
  \item{id}{Language id}
  \item{code}{ISO code}
  \item{type}{Language or dialect}
  \item{name}{Language name}
  \item{latitude}{latitude}
  \item{longitude}{longitude}
  \item{gen1}{Language Family}
  \item{gen2}{Language Family}
  \item{tones}{Inventory of tones}
  \item{syllab}{Syllab structure}
  \item{cluster}{Cluster}
  \item{finals}{Finals}
  \item{source}{Source}
  \item{comment}{Comment}
  \item{Contr}{Contributor}
  \item{Segment type}{Vowels or consonants}
  \item{segments}{Segments}
  \item{glottocode}{Glottocode}
}
}
\keyword{datasets}
