% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottolog.R
\docType{data}
\name{glottolog}
\alias{glottolog}
\title{Catalogue of languages of the world}
\format{A data frame with 8286 rows and 7 variables:
\describe{
  \item{iso}{code based on ISO 639-3 \url{http://www-01.sil.org/iso639-3/}}
  \item{languoid}{name of the languoid}
  \item{affiliation}{genealogical affiliation}
  \item{macro_area}{have six values Africa, Australia, Eurasia, North America, Papunesia, South America}
  \item{country}{list of countries, where the language is spoken}
  \item{latitude}{latitude}
  \item{longitude}{longitude}
}}
\source{
\url{http://glottolog.org/}
}
\usage{
glottolog
}
\description{
A dataset containes the catalogue of languages of the world
involving genealogical affiliation, macro-area, country, iso code,
and coordinates.
}
\details{
Glottolog 2.7. Hammarstrom, Harald & Forkel, Robert & Haspelmath, Martin &
Bank, Sebastian. 2016.
Max Planck Institute for the Science of Human History.
Accessed on 2016-06-15.
}
\keyword{datasets}

