% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltc.iso.R
\name{gltc.iso}
\alias{gltc.iso}
\title{Get Glottocode by ISO 639--3 code}
\usage{
gltc.iso(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the Glottocodes.}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of ISO 639--3 codes and returns Glottocodes.
}
\examples{
gltc.iso('ady')
gltc.iso(c('ady', 'rus'))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

