% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltc.lang.R
\name{gltc.lang}
\alias{gltc.lang}
\title{Get Glottocode by language}
\usage{
gltc.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns Glottocode.
}
\examples{
gltc.lang('Adyghe')
gltc.lang(c('Adyghe', 'Udi'))
}
\seealso{
\code{\link{aff.lang.R}}, \code{\link{area.lang.R}}, \code{\link{country.lang.R}}, \code{\link{iso.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{subc.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
