% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grambank.feature.R
\name{grambank.feature}
\alias{grambank.feature}
\title{Download Grambank data}
\usage{
grambank.feature(features, na.rm = TRUE)
}
\arguments{
\item{features}{A character vector that define with a feature ids from Grambank (e. g. "gb026", "gb042").}

\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}
}
\description{
This function downloads data from Grambank (\url{https://grambank.clld.org/}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# grambank.feature(c("gb026", "gb042"))
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{bivaltyp.feature}}, \code{\link{eurasianphonology.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{vanuatu.feature}}, \code{\link{wals.feature}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
