% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.feature.R
\name{ids.feature}
\alias{ids.feature}
\title{Download data from The Intercontinental Dictionary Series}
\usage{
ids.feature(features = c(), na.rm = TRUE)
}
\arguments{
\item{features}{A vector with parameters from conepts (\url{https://ids.clld.org/parameters}))}

\item{na.rm}{Logical. If TRUE function removes all languages not available in the glottolog database available in lingtypology. By default is TRUE.}
}
\description{
This function downloads data from The Intercontinental Dictionary Series (\url{https://ids.clld.org/}). You need the internet connection.
}
\author{
Mikhail Leonov
}
