% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.gltc.R
\name{iso.gltc}
\alias{iso.gltc}
\title{Get ISO 639--3 code by Glottocode}
\usage{
iso.gltc(x)
}
\arguments{
\item{x}{A character vector of Glottocodes.}
}
\description{
Takes any vector of Glotocodes and returns ISO code.
}
\examples{
iso.gltc('adyg1241')
iso.gltc(c('adyg1241', 'udii1243'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
