% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lang.R
\name{iso.lang}
\alias{iso.lang}
\title{Get ISO 639--3 code by language}
\usage{
iso.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns ISO code.
}
\examples{
iso.lang('Adyghe')
iso.lang(c('Adyghe', 'Udi'))
}
\seealso{
\code{\link{aff.lang.R}}, \code{\link{area.lang.R}}, \code{\link{country.lang.R}}, \code{\link{gltc.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{subc.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
