% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3.iso1.R
\name{iso3.iso1}
\alias{iso3.iso1}
\title{Get ISO 639-3 code from ISO 639-1}
\usage{
iso3.iso1(x)
}
\arguments{
\item{x}{A character vector of ISO 639-3 codes.}
}
\description{
Takes any vector of ISO 639-1 codes and returns ISO 639-3 code.
}
\examples{
iso3.iso1('bs')
iso3.iso1(c('co', 'it', 'ar'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}
\author{
Ekaterina Zalivina <zalivina01@mail.ru>
}
