% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.aff.R
\name{lang.aff}
\alias{lang.aff}
\title{Get languoids by affiliation}
\usage{
lang.aff(x, list = FALSE)
}
\arguments{
\item{x}{A character vector of the affiliations (can be written in lower case)}

\item{list}{logical. If TRUE, returns a list of languoids, if ELSE return a named vector.}
}
\description{
Takes any vector of affiliations and return languoids.
}
\examples{
lang.aff("Balto-Slavic")
lang.aff(c("East Slavic", "West Slavic"))
}
\author{
George Moroz <agricolamz@gmail.com>
}

