% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.gltc.R
\name{lang.gltc}
\alias{lang.gltc}
\title{Get language by Glottocode}
\usage{
lang.gltc(x)
}
\arguments{
\item{x}{A character vector of the Glottocodes.}
}
\description{
Takes any vector of Glottocodes and return languages.
}
\examples{
lang.gltc('adyg1241')
lang.gltc(c('adyg1241', 'udii1243'))
}
\seealso{
\code{\link{lang.aff}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
