% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat.lang.R
\name{lat.lang}
\alias{lat.lang}
\title{Get latitude by languoid}
\usage{
lat.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return latitude.
}
\author{
George Moroz <agricolamz@gmail.com>
}

