% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level.lang.R
\name{level.lang}
\alias{level.lang}
\title{Get a level of language by language}
\usage{
level.lang(x)
}
\arguments{
\item{x}{character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns a level of language.
}
\examples{
level.lang('Russian Sign Language')
level.lang(c('Archi', 'Chechen'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{country.lang}}, \code{\link{gltc.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}, \code{\link{subc.lang}}, \code{\link{url.lang}}
}
\author{
Sasha Shakhnova
}
