% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long.lang.R
\name{long.lang}
\alias{long.lang}
\title{Get longitude by languoid}
\usage{
long.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return longitude.
}
\author{
George Moroz <agricolamz@gmail.com>
}

