% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.url.R
\name{make.url}
\alias{make.url}
\title{Make a url-link to glottolog page for a language}
\usage{
make.url(x, popup = "", glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of languages (can be written in lower case)}

\item{popup}{character vector of strings that will appear in pop-up window of the function map.feature}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of languages and return links to glottolog pages.
}
\examples{
make.url('Korean')
make.url(c('Gangou', 'Hachijo', 'Adyghe', 'Ganai'))
}
\author{
George Moroz <agricolamz@gmail.com>
}

