% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.feature.R
\name{map.feature}
\alias{map.feature}
\title{Map a set of linguoids and color them by feature}
\usage{
map.feature(languages, features = "none", popup = "", title = NULL)
}
\arguments{
\item{x}{A character vector of linguoids (can be written in lower case)}
}
\description{
Takes any vector of linguoids and return a map.
}
\examples{
map.lang(c("Adyghe", "Russian"))
## Map all Slavic languages
map.lang(lang.aff(c("Slavic")))
}
\author{
George Moroz <agricolamz@gmail.com>
}

