% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.iso.R
\name{map.iso}
\alias{map.iso}
\title{Map a set of linguoids}
\usage{
map.iso(x)
}
\arguments{
\item{x}{A character vector of ISO codes}
}
\description{
Takes any vector of ISO codes and return a map.
}
\examples{
map.iso(c("ady", "rus"))
}
\author{
George Moroz <agricolamz@gmail.com>
}

