% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_mangueanIC.feature.R
\name{oto_mangueanIC.feature}
\alias{oto_mangueanIC.feature}
\title{Download Oto-Manguean Inflectional Class Database data}
\usage{
oto_mangueanIC.feature()
}
\description{
This function downloads data from Oto-Manguean Inflectional Class Database (\url{https://oto-manguean.surrey.ac.uk/}) and creates a language column with the names from lingtypology database. You need the internet connection.
}
\examples{
# oto_mangueanIC.feature()
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
