% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oto_mangueanIC.R
\docType{data}
\name{oto_mangueanIC}
\alias{oto_mangueanIC}
\title{Oto-Manguean Inflectional Class Database Language identifiers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 2 columns.
}
\usage{
oto_mangueanIC
}
\description{
Language identifiers from Oto-Manguean Inflectional Class Database (\url{https://oto-manguean.surrey.ac.uk/}). This dataset is created for \code{\link{oto_mangueanIC.feature}} function.
}
\details{
#' @format A data frame with 20 rows and 2 variables:
\describe{
  \item{Language.name}{Languaeg names from Oto-Manguean Inflectional Class Database}
  \item{language}{Language names from Glottolog database}
}
}
\keyword{datasets}
