% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoible.R
\docType{data}
\name{phoible}
\alias{phoible}
\title{Phoible glottolog - language correspondencies}
\format{
A data frame with 2185 rows and 2 variables:
\describe{
  \item{language}{language}
  \item{Glottocode}{Glottocode}
}
}
\usage{
phoible
}
\description{
Language correspondencies for Phoible (\url{https://phoible.org/}). This dataset is created for \code{\link{phoible.feature}} function.
}
\keyword{datasets}
