% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonological_profiles.R
\docType{data}
\name{phonological_profiles}
\alias{phonological_profiles}
\title{Number of consonants and presence of ejectives}
\format{
A data frame with 19 rows and 4 variables:
\describe{
  \item{language}{language name}
  \item{consonants}{number of consonants. Based on UPSID database.}
  \item{vowels}{number of vowels. Based on UPSID database.}
  \item{ejectives}{presence of ejective sounds.}
  \item{tone}{presence of tone.}
  \item{stress}{presence of stress.}
  \item{long_vowels}{presence of long vowels.}
}
}
\usage{
phonological_profiles
}
\description{
Number of consonants and presence of ejectives
}
\keyword{datasets}
