% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.points.R
\name{polygon.points}
\alias{polygon.points}
\title{Get kernel density estimation poligon from coordinates}
\usage{
polygon.points(latitude, longitude)
}
\arguments{
\item{latitude}{numeric vector of latitudes}

\item{longitude}{numeric vector of longitudes}
}
\description{
This function is based on this answer: https://gis.stackexchange.com/a/203623
}
