% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/providers.R
\docType{data}
\name{providers}
\alias{providers}
\title{Providers}
\format{
A list of characters
}
\source{
\url{https://leaflet-extras.github.io/leaflet-providers/preview/}
}
\usage{
providers
}
\description{
List of all providers with their variations taken from leaflet package
}
\keyword{datasets}
