% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segbo.feature.R
\name{segbo.feature}
\alias{segbo.feature}
\title{Download SegBo data}
\usage{
segbo.feature(languages)
}
\arguments{
\item{languages}{vector of Glotocodes. Use 'all' to download all data.}
}
\description{
This function downloads data from SegBo (\url{https://github.com/segbo-db/segbo}). You need the internet connection.
}
\examples{
# segbo.feature('abau1245')
# segbo.feature('abui1241', 'aleu1260')
# segbo.feature('all')
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{bivaltyp.feature}}, \code{\link{eurasianphonology.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{vanuatu.feature}}, \code{\link{wals.feature}}

\code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Ekaterina Zalivina <zalivina01@mail.ru>
}
