% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcomparisons.feature.R
\name{soundcomparisons.feature}
\alias{soundcomparisons.feature}
\title{Download SOUNDCOMPARISONS data}
\usage{
soundcomparisons.feature(word)
}
\arguments{
\item{word}{A character vector that define with a feature ids from SOUNDCOMPARISONS (e. g. "one", "sharp_fem", "near_neut", "on_the_left", "    I_will_give", "write_ipv_sg", "your_pl_pl").}

\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}
}
\description{
This function downloads data from SOUNDCOMPARISONS (\url{https://soundcomparisons.com/}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
## Not run:
soundcomparisons.feature(c("sun", "house"))
## End(Not run)
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}
}
\author{
Anna Smirnova <annedadaa@gmail.com>
}
