% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundcomparisons.R
\docType{data}
\name{soundcomparisons}
\alias{soundcomparisons}
\title{SOUNDCOMPARISONS's Language identifiers}
\format{
An object of class \code{data.frame} with 556 rows and 3 columns.
}
\usage{
soundcomparisons
}
\description{
Language identifiers from SOUNDCOMPARISONS (\url{https://soundcomparisons.com/}). This dataset is created for \code{\link{soundcomparisons.feature}} function.
}
\details{
#' @format A data frame with 556 rows and 2 variables:
\describe{
  \item{LanguageName}{SOUNDCOMPARISONS language identifier}
  \item{LanduageId}{Language Id}
}
}
\keyword{datasets}
