% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subc.lang.R
\name{subc.lang}
\alias{subc.lang}
\title{Get subclassification by language}
\usage{
subc.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and returns subclassification in the Newick tree format.
}
\examples{
subc.lang('Korean')
subc.lang(c('Korean', 'Lechitic'))
}
\seealso{
\code{\link{aff.lang.R}}, \code{\link{area.lang.R}}, \code{\link{country.lang.R}}, \code{\link{gltc.lang.R}}, \code{\link{iso.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
