% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uralex.feature.R
\name{uralex.feature}
\alias{uralex.feature}
\title{Download UraLex data}
\usage{
uralex.feature(na.rm = TRUE)
}
\arguments{
\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}
}
\description{
This function downloads data from UraLex (https://github.com/lexibank/uralex/) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# uralex.feature()
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
