% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uratyp.feature.R
\name{uratyp.feature}
\alias{uratyp.feature}
\title{Download Uratyp data}
\usage{
uratyp.feature(languages, type = "GB")
}
\arguments{
\item{languages}{language names from data. Possible values: 'Soikkola_Ingrian', 'West_Votic', 'Voro_South_Estonian', 'Livvi_Karelian', 'Courland_Livonian', 'Ludian', 'Central_Veps', 'North_Karelian', 'Tver_Karelian', 'Estonian', 'Finnish', 'Pite_Saami', 'Inari_Saami', 'Lule_Saami', 'South_Saami', 'North_Saami', 'Skolt_Saami', 'Meadow_Mari', 'Hill_Mari', 'Erzya', 'Moksha', 'Komi_Zyrian', 'Udmurt', 'Hungarian', 'East_Mansi', 'North_Mansi', 'Kazym_Khanty', 'Surgut_Khanty', 'Forest_Enets', 'Tundra_Nenets', 'South_Selkup', 'Kamas', 'Nganasan', 'Komi_Permyak', 'Ume_Saami'.}

\item{type}{type of list of features. Possible values: 'GB', 'UT'. By default is 'GB'. More about types (\url{https://uralic.clld.org}).}
}
\description{
This function downloads data from Uratyp (\url{https://uralic.clld.org}). You need the internet connection.
}
\examples{
# uratyp.feature('Central_Veps')
# uratyp.feature('Hungarian', 'UT')
# uratyp.feature(c('Inari_Saami', 'Erzya'), 'UT')
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{bivaltyp.feature}}, \code{\link{eurasianphonology.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{vanuatu.feature}}, \code{\link{wals.feature}}

\code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Ekaterina Zalivina <zalivina01@mail.ru>
}
