% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valpal.feature.R
\name{valpal.feature}
\alias{valpal.feature}
\title{Download ValPaL data}
\usage{
valpal.feature(na.rm = FALSE)
}
\arguments{
\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is FALSE.}
}
\description{
This function downloads data from ValPal (http://www.valpal.info/) and changes language names to the names from lingtypology database. You need the internet connection.
}
\examples{
# valpal.feature()
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{wals.feature}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
