% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanuatu.feature.R
\name{vanuatu.feature}
\alias{vanuatu.feature}
\title{Download Vanuatu Voices data}
\usage{
vanuatu.feature(features, na.rm = TRUE)
}
\arguments{
\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}

\item{feature}{A string with a parameter from Conepts (\url{https://vanuatuvoices.clld.org/parameters}))}
}
\description{
This function downloads data from Vanuatu Voices (\url{https://vanuatuvoices.clld.org/}). You need the internet connection.
}
\author{
Mikhail Leonov
}
