% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\docType{data}
\name{wals}
\alias{wals}
\title{WALS's Language identifiers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2679 rows and 2 columns.}
\usage{
wals
}
\description{
Language identifiers from WALS (\url{http://wals.info/}). This dataset is created for \code{\link{wals.feature}} function.
}
\details{
#' @format A data frame with 2950 rows and 2 variables:
\describe{
  \item{wals code}{WALS language identifier}
  \item{glottocode}{Glottocode}
}
}
\keyword{datasets}
