% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcsfeature.R
\name{wcs.feature}
\alias{wcs.feature}
\title{Downloads data from World Color Survey Data Archives.}
\usage{
wcs.feature()
}
\description{
This function downloads data from World Color Survey Data Archives. (\url{http://www1.icsi.berkeley.edu/wcs/data.html}).
}
\examples{

wcs.feature()

}
\seealso{

}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
