% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs.R
\docType{data}
\name{wcs}
\alias{wcs}
\title{WCS data}
\format{

}
\usage{
wcs
}
\description{
Data from The database of WCS Data Archives (\url{https://www1.icsi.berkeley.edu/wcs/data.html}). This dataset is created for \code{\link{wcs.feature}} function.
}
\keyword{datasets}
