% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.lang.R
\name{aff.lang}
\alias{aff.lang}
\title{Get affiliation by language}
\usage{
aff.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns affiliation.
}
\examples{
aff.lang('Korean')
aff.lang(c('Korean', 'Polish'))
}
\seealso{
\code{\link{area.lang.R}}, \code{\link{country.lang.R}}, \code{\link{gltc.lang.R}}, \code{\link{iso.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{subc.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
