% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.lang.R
\name{aff.lang}
\alias{aff.lang}
\title{Get affiliation by language}
\usage{
aff.lang(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of languages and return affiliation.
}
\examples{
aff.lang('Korean')
aff.lang(c('Korean', 'Polish'))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

