% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aff.lang.R
\name{aff.lang}
\alias{aff.lang}
\title{Get affiliation by languoid}
\usage{
aff.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return affiliation.
}
\examples{
aff.lang("Korean")
aff.lang(c("Korean", "Polish"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

