% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.lang.R
\name{area.lang}
\alias{area.lang}
\title{Get macro area by language}
\usage{
area.lang(x)
}
\arguments{
\item{x}{character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns macro area.
}
\examples{
area.lang('Adyghe')
area.lang(c('Adyghe', 'Aduge'))
}
\seealso{
\code{\link{aff.lang.R}}, \code{\link{country.lang.R}}, \code{\link{gltc.lang.R}}, \code{\link{iso.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{subc.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
