% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.lang.R
\name{area.lang}
\alias{area.lang}
\title{Get macro area by languoid}
\usage{
area.lang(x)
}
\arguments{
\item{x}{character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return macro area.
}
\examples{
area.lang("Adyghe")
area.lang(c("Adyghe", "Aduge"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{aff.lang}}, \code{\link{country.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

