% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotyp.R
\docType{data}
\name{autotyp}
\alias{autotyp}
\title{AUTOTYP's Language identifiers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1336 rows and 3 columns.
}
\usage{
autotyp
}
\description{
Language identifiers from AUTOTYP v. 0.1.4 (\url{https://github.com/autotyp/autotyp-data/}). This dataset is created for \code{\link{autotyp.feature}} function.
}
\details{
#' @format A data frame with 1336 rows and 3 variables:
\describe{
  \item{path}{path to the dataset in autotyp repo}
  \item{variable}{variable name}
  \item{file}{topic name}
}
}
\keyword{datasets}
