% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bantu.R
\docType{data}
\name{bantu}
\alias{bantu}
\title{BANTU's Language identifiers}
\format{
A data frame with 430 rows and 2 variables:
\describe{
  \item{id}{BANTU word id}
  \item{word}{word}
}
}
\usage{
bantu
}
\description{
Language identifiers from BANTU (\url{https://abvd.shh.mpg.de/bantu/index.php}). This dataset is created for \code{\link{bantu.feature}} function.
}
\keyword{datasets}
