% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivaltyp.feature.R
\name{bivaltyp.feature}
\alias{bivaltyp.feature}
\title{Download BivalTyp data}
\usage{
bivaltyp.feature()
}
\description{
This function downloads data from BivalTyp (\url{https://www.bivaltyp.info/}) and changes language names to the names from lingtypology database. You need the internet connection.
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}

\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{eurasianphonology.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{vanuatu.feature}}, \code{\link{wals.feature}}
# bivaltyp.feature()
}
\author{
George Moroz <agricolamz@gmail.com>
}
