% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Catalogue of countries names.}
\format{A data frame with 86 rows and 3 variables:
\describe{
  \item{common}{common name}
  \item{official}{official name}
  \item{abbreviation}{abreviated name}
  \item{official_languages}{official languages from the given country}
}}
\usage{
countries
}
\description{
Catalogue of countries names.
}
\keyword{datasets}

