% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.lang.R
\name{country.lang}
\alias{country.lang}
\title{Get subclassification by language}
\usage{
country.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns countries where those languages are used as ISO 3166-1 alpha-2 codes.
}
\examples{
country.lang('Korean')
subc.lang(c('Korean', 'Lechitic'))
}
\seealso{
\code{\link{aff.lang.R}}, \code{\link{area.lang.R}}, \code{\link{gltc.lang.R}}, \code{\link{iso.lang.R}}, \code{\link{lat.lang.R}}, \code{\link{long.lang.R}}, \code{\link{subc.lang.R}}, \code{\link{url.lang.R}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
