% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.lang.R
\name{country.lang}
\alias{country.lang}
\title{Get country by language}
\usage{
country.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns countries where those languages are used as ISO 3166-1 alpha-2 codes.
}
\examples{
country.lang('Korean')
country.lang(c('Korean', 'Polish'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{gltc.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}, \code{\link{subc.lang}}, \code{\link{url.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
