% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejective_and_n_consonants.R
\docType{data}
\name{ejective_and_n_consonants}
\alias{ejective_and_n_consonants}
\title{Number of consonants and presence of ejectives}
\format{A data frame with 27 rows and 3 variables:
\describe{
  \item{language}{language name}
  \item{n.cons.lapsyd}{number of consonants. Based on \href{http://www.lapsyd.ddl.ish-lyon.cnrs.fr/lapsyd/}{LAPSyD} database.}
  \item{ejectives}{presence of ejective sounds}
}}
\usage{
ejective_and_n_consonants
}
\description{
Number of consonants and presence of ejectives
}
\keyword{datasets}

