% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eurasianphonology.feature.R
\name{eurasianphonology.feature}
\alias{eurasianphonology.feature}
\title{Opens data from the database of Eurasian phonological inventories}
\usage{
eurasianphonology.feature()
}
\description{
This function opens downloaded data from the database of Eurasian phonological inventories (\url{https://eurphon.info}).
}
\examples{

eurasianphonology.feature()

}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{bivaltyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{vanuatu.feature}}, \code{\link{wals.feature}}
}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
