% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gltc.lang.R
\name{gltc.lang}
\alias{gltc.lang}
\title{Get Glottocode by language}
\usage{
gltc.lang(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of languages and returns Glottocode.
}
\examples{
gltc.lang('Adyghe')
gltc.lang(c('Adyghe', 'Udi'))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

