% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.glottolog.R
\name{is.glottolog}
\alias{is.glottolog}
\title{Is it in glottolog?}
\usage{
is.glottolog(x, response = F)
}
\arguments{
\item{x}{A character vector of linguoids (can be written in lower case)or ISO codes}
}
\description{
Takes any vector of linguoids or ISO codes and return a logical vector.
}
\author{
George Moroz <agricolamz@gmail.com>
}

