% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lang.R
\name{iso.lang}
\alias{iso.lang}
\title{Get ISO 639--3 code by language}
\usage{
iso.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns ISO code.
}
\examples{
iso.lang('Kabardian')
iso.lang(c('Kabardian', 'Udi'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{gltc.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}, \code{\link{subc.lang}}, \code{\link{url.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
