% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lang.R
\name{iso.lang}
\alias{iso.lang}
\title{Get ISO 639--3 code by language}
\usage{
iso.lang(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of languages and returns ISO code.
}
\examples{
iso.lang('Adyghe')
iso.lang(c('Adyghe', 'Udi'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
