% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso.lang.R
\name{iso.lang}
\alias{iso.lang}
\title{Get ISO 639-3 code by languoid}
\usage{
iso.lang(x)
}
\arguments{
\item{x}{A character vector of the languoids (can be written in lower case)}
}
\description{
Takes any vector of languoids and return ISO code.
}
\examples{
iso.lang("Adyghe")
iso.lang(c("Adyghe", "Udi"))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}
}

