% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.aff.R
\name{lang.aff}
\alias{lang.aff}
\title{Get languoids by affiliation}
\usage{
lang.aff(x)
}
\arguments{
\item{x}{A character vector of the affiliations (can be written in lower case)}
}
\description{
Takes any vector of affiliations and return languoids.
}
\author{
George Moroz <agricolamz@gmail.com>
}

