% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.aff.R
\name{lang.aff}
\alias{lang.aff}
\title{Get languages by affiliation}
\usage{
lang.aff(x, list = FALSE)
}
\arguments{
\item{x}{A character vector of the affiliations (can be written in lower case)}

\item{list}{logical. If TRUE, returns a list of languages, if FALSE return a named vector.}
}
\description{
Takes any vector of affiliations and return languages.
}
\examples{
lang.aff('Slavic')
lang.aff(c('Slavic', 'Celtic'))
lang.aff(c('Slavic', 'Celtic'), list = TRUE)
}
\seealso{
\code{\link{lang.iso}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
