% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.country.R
\name{lang.country}
\alias{lang.country}
\title{Get language by country}
\usage{
lang.country(x)
}
\arguments{
\item{x}{character vector of the countries (in alpha-2 ISO codes)}
}
\description{
Takes any vector of countries and returns languages.
}
\examples{
lang.country('AD')
lang.country(c('AD', 'AE'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{country.lang}}, \code{\link{gltc.lang}}, \code{\link{iso.lang}}, \code{\link{lat.lang}}, \code{\link{long.lang}}, \code{\link{subc.lang}}, \code{\link{url.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
