% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.country.R
\name{lang.country}
\alias{lang.country}
\title{Get languoids by country}
\usage{
lang.country(x, list = FALSE)
}
\arguments{
\item{x}{character vector of the countries (can be written in lower case)}

\item{list}{logical. If TRUE, returns a list of languoids, if FALSE return a vector.}
}
\description{
Takes any vector of countries and return languoids.
}
\examples{
lang.country("Bali")
lang.country(c("Bali", "Luxembourg"))
lang.country(c("Bali", "Luxembourg"), list = TRUE)
## What languoids are both in North Korea and in South Korea?
lang.country("Korea")
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{lang.aff}}, \code{\link{lang.iso}}
}

