% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.gltc.R
\name{lang.gltc}
\alias{lang.gltc}
\title{Get language by Glottocode}
\usage{
lang.gltc(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the Glottocodes.}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of Glottocodes and return languages.
}
\examples{
lang.gltc('adyg1241')
lang.gltc(c('adyg1241', 'udii1243'))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{lang.aff}}, \code{\link{lang.country}}
}

