% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.iso.R
\name{lang.iso}
\alias{lang.iso}
\title{Get language by ISO 639--3 code}
\usage{
lang.iso(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the ISO codes.}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of ISO codes and return languages.
}
\examples{
lang.iso('ady')
lang.iso(c('ady', 'rus'))
}
\seealso{
\code{\link{lang.aff}}, \code{\link{lang.country}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
