% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.iso.R
\name{lang.iso}
\alias{lang.iso}
\title{Get language by ISO 639--3 code}
\usage{
lang.iso(x)
}
\arguments{
\item{x}{A character vector of the ISO codes.}
}
\description{
Takes any vector of ISO codes and return languages.
}
\examples{
lang.iso('ady')
lang.iso(c('ady', 'rus'))
}
\seealso{
\code{\link{lang.aff}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
