% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat.lang.R
\name{lat.lang}
\alias{lat.lang}
\title{Get latitude by language}
\usage{
lat.lang(x)
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}
}
\description{
Takes any vector of languages and returns latitude.
}
\examples{
lat.lang('Adyghe')
long.lang('Adyghe')
lat.lang(c('Adyghe', 'Russian'))
long.lang(c('Adyghe', 'Russian'))
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{gltc.lang}}, \code{\link{iso.lang}}, \code{\link{long.lang}}, \code{\link{subc.lang}}, \code{\link{url.lang}}
}
\author{
George Moroz <agricolamz@gmail.com>
}
