% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lat.lang.R
\name{lat.lang}
\alias{lat.lang}
\title{Get latitude by language}
\usage{
lat.lang(x, glottolog.source = "modified")
}
\arguments{
\item{x}{A character vector of the languages (can be written in lower case)}

\item{glottolog.source}{A character vector that define which glottolog database is used: 'original' or 'modified' (by default)}
}
\description{
Takes any vector of languages and return latitude.
}
\examples{
lat.lang('Adyghe')
long.lang('Adyghe')
lat.lang(c('Adyghe', 'Russian'))
long.lang(c('Adyghe', 'Russian'))
}
\author{
George Moroz <agricolamz@gmail.com>
}
\seealso{
\code{\link{aff.lang}}, \code{\link{area.lang}}, \code{\link{country.lang}}, \code{\link{iso.lang}}, \code{\link{long.lang}}
}

