% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.feature.R
\name{map.feature}
\alias{map.feature}
\title{Map a set of linguoids and color them by feature}
\usage{
map.feature(languages, features = "none", stroke.features = NULL,
  popup = "", latitude = NULL, longitude = NULL, color = NULL,
  stroke.color = NULL, title = NULL, control = TRUE, ...)
}
\arguments{
\item{languages}{character vector of linguoids (can be written in lower case)}

\item{features}{character vector of features}

\item{stroke.features}{additional independent stroke features}

\item{popup}{character vector of strings that will appear in pop-up window}

\item{latitude}{numeric vector of latitudes}

\item{longitude}{numeric vector of longitudes}

\item{color}{vector of colors}

\item{stroke.color}{vector of stroke colors}

\item{title}{of a legend}

\item{control}{logical. If FALSE, function doesn't show layer control buttons.}

\item{...}{further arguments of leaflet package.}
}
\description{
Takes any vector of linguoids and return a map.
}
\examples{
map.feature(c("Adyghe", "Russian"))

## Map all Slavic languages
map.feature(lang.aff(c("Slavic")))

## Color linguoids by feature
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"))
map.feature(df$lang, df$feature)

## Adding pop-up
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Adyghe", "Adyghe", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup)

## Adding title
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Adyghe", "Adyghe", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup, title = "type of a language")

## Add your own coordinates
map.feature("Adyghe", latitude = 43, longitude = 57)

## Add you own colors
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Adyghe", "Adyghe", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup, color = c("green", "navy"))

## Remove control buttons
map.feature(lang.aff("Sign"), control = FALSE)

## Use strokes for aditional features
map.feature(df$language, features = df$languoid, stroke.features = df$language, latitude = df$latitude, longitude = df$longitude, control = F)
}
\author{
George Moroz <agricolamz@gmail.com>
}

