% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.feature.R
\name{map.feature}
\alias{map.feature}
\title{Map a set of linguoids and color them by feature}
\usage{
map.feature(languages, features = "none", popup = "", latitude = NULL,
  longitude = NULL, color = NULL, title = NULL)
}
\arguments{
\item{languages}{character vector of linguoids (can be written in lower case)}

\item{features}{character vector of features}

\item{popup}{character vector of strings that will appear in pop-up window}

\item{latitude}{numeric vector of latitudes}

\item{longitude}{numeric vector of longitudes}

\item{color}{vector of colors}

\item{title}{of a legend}
}
\description{
Takes any vector of linguoids and return a map.
}
\examples{
map.feature(c("Adyghe", "Russian"))

## Map all Slavic languages
map.feature(lang.aff(c("Slavic")))

## Color linguoids by feature
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"))
map.feature(df$lang, df$feature)

## Adding pop-up
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Adyghe", "Adyghe", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup)

## Adding title
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Adyghe", "Adyghe", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup, title = "type of a language")

## Add your own coordinates
map.feature("Adyghe", latitude = 43, longitude = 57)
}
\author{
George Moroz <agricolamz@gmail.com>
}

