% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.feature.R
\name{map.feature}
\alias{map.feature}
\title{Create a map}
\usage{
map.feature(languages, features = "none", popup = "", label = "",
  label.hide = FALSE, label.fsize = 14, label.position = "right",
  stroke.features = NULL, latitude = NULL, longitude = NULL,
  color = NULL, stroke.color = NULL, image.url = NULL,
  image.width = 100, image.height = 100, image.X.shift = 0,
  image.Y.shift = 0, title = NULL, stroke.title = NULL, control = FALSE,
  legend = TRUE, legend.opacity = 1, legend.position = "topright",
  stroke.legend = TRUE, stroke.legend.opacity = 1,
  stroke.legend.position = "bottomleft", radius = 5, stroke.radius = 9.5,
  opacity = 1, stroke.opacity = 1, scale.bar = TRUE,
  scale.bar.position = "bottomleft", minimap = FALSE,
  minimap.position = "bottomright", minimap.width = 150,
  minimap.height = 150, tile = "OpenStreetMap.Mapnik", tile.name = NULL,
  glottolog.source = "modified")
}
\arguments{
\item{languages}{character vector of linguoids (can be written in lower case)}

\item{features}{character vector of features}

\item{popup}{character vector of strings that will appear in pop-up window}

\item{label}{character vector of strings that will appear near points}

\item{label.hide}{logical. If FALSE, labels are displayed allways. If TRUE, labels are displayed on mouse over. By default is TRUE.}

\item{label.fsize}{numeric value of the label font size. By default is 14.}

\item{label.position}{the position of labels: "left", "right", "top", "bottom"}

\item{stroke.features}{additional independent stroke features}

\item{latitude}{numeric vector of latitudes}

\item{longitude}{numeric vector of longitudes}

\item{color}{vector of colors}

\item{stroke.color}{vector of stroke colors}

\item{image.url}{character vector of URLs with an images}

\item{image.width}{numeric vector of image widths}

\item{image.height}{numeric vector of image heights}

\item{image.X.shift}{numeric vector of image's X axis shift relative to the latitude-longitude point}

\item{image.Y.shift}{numeric vector of image's Y axis shift relative to the latitude-longitude point}

\item{title}{title of a legend}

\item{stroke.title}{title of a stroke-feature legend}

\item{control}{logical. If TRUE, function show layer control buttons. By default is TRUE.}

\item{legend}{logical. If TRUE, function show legend. By default is FALSE.}

\item{legend.opacity}{a numeric vector of legend opacity.}

\item{legend.position}{the position of the legend: "topright", "bottomright", "bottomleft","topleft"}

\item{stroke.legend}{logical. If TRUE, function show stroke.legend. By default is FALSE.}

\item{stroke.legend.opacity}{a numeric vector of stroke.legend opacity.}

\item{stroke.legend.position}{the position of the stroke.legend: "topright", "bottomright", "bottomleft","topleft"}

\item{radius}{a numeric vector of radii for the circles.}

\item{stroke.radius}{a numeric vector of stroke radii for the circles.}

\item{opacity}{a numeric vector of marker opacity.}

\item{stroke.opacity}{a numeric vector of stroke opacity.}

\item{scale.bar}{logical. If TRUE, function show scale-bar. By default is TRUE.}

\item{scale.bar.position}{the position of the scale-bar: "topright", "bottomright", "bottomleft","topleft"}

\item{minimap}{= FALSE,}

\item{minimap.position}{the position of the minimap: "topright", "bottomright", "bottomleft","topleft"}

\item{minimap.width}{The width of the minimap in pixels.}

\item{minimap.height}{The height of the minimap in pixels.}

\item{tile}{a character verctor with a map tiles, popularized by Google Maps. See \href{https://leaflet-extras.github.io/leaflet-providers/preview/index.html}{here} for the complete set.}

\item{tile.name}{a character verctor with a user's map tiles' names}

\item{glottolog.source}{A character vector that define which glottolog database is used: "original" or "modified" (by default)}
}
\description{
Map a set of linguoids and color them by feature or two sets of features.
}
\examples{
map.feature(c("Adyghe", "Russian"))

## All Sign languages
map.feature(lang.aff("Sign"))

## Map all Slavic languages
map.feature(lang.aff(c("Slavic")))

## Add control buttons
map.feature(c("Adyghe", "Russian"), control = TRUE)

## Color linguoids by feature
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"))
map.feature(df$lang, df$feature)
## ... or add a control buttons for features
map.feature(df$lang, df$feature, control = TRUE)

## Adding pop-up
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Circassian", "Circassian", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup)

## Adding labels
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Circassian", "Circassian", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, label = df$lang)

## Adding title
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Circassian", "Circassian", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup, title = "type of a language")

## Add your own coordinates
map.feature("Adyghe", latitude = 43, longitude = 57)

## Change map tile
map.feature("Adyghe", tile = "Thunderforest.OpenCycleMap")
map.feature("Adyghe", tile = c("OpenStreetMap.BlackAndWhite", "Thunderforest.OpenCycleMap"))
map.feature("Adyghe", tile = "Thunderforest.OpenCycleMap", tile.name = "colored")

## Add you own colors
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Circassian", "Circassian", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup, color = c("green", "navy"))

## Map two sets of features
df <- data.frame(lang = c("Adyghe", "Kabardian", "Polish", "Russian", "Bulgarian"),
feature = c("polysynthetic", "polysynthetic", "fusion", "fusion", "fusion"),
popup = c("Circassian", "Circassian", "Slavic", "Slavic", "Slavic"))
map.feature(df$lang, df$feature, df$popup,
stroke.features = df$popup)

## Add a pictures to plot
df <- data.frame(lang = c("Russian", "Russian"),
lat  = c(55.75, 59.95),
long = c(37.616667, 30.3),
urls = c("https://goo.gl/5OUv1E", "https://goo.gl/UWmvDw"))
map.feature(languages = df$lang,
latitude = df$lat,
longitude = df$long,
image.url = df$urls)

## Add a minimap to plot
map.feature(c("Adyghe", "Russian"), minimap = TRUE)

## Remove scale bar
map.feature(c("Adyghe", "Russian"), scale.bar = FALSE)

}
\author{
George Moroz <agricolamz@gmail.com>
}

