% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uralex.R
\docType{data}
\name{uralex}
\alias{uralex}
\title{UraLex's Language identifiers}
\format{A data frame with 27 rows and 3 variables:
\describe{
  \item{language}{language name from database}
  \item{Glottocode}{Glottocodes}
  \item{language2}{language from lingtypology}
}}
\usage{
uralex
}
\description{
Language identifiers from UraLex (\url{https://github.com/lexibank/uralex}). This dataset is created for \code{\link{uralex.feature}} function.
}
\keyword{datasets}
