% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanuatu.feature.R
\name{vanuatu.feature}
\alias{vanuatu.feature}
\title{Download Vanuatu Voices data}
\usage{
vanuatu.feature(feature)
}
\arguments{
\item{feature}{A string with a parameter from Conepts (\url{https://vanuatuvoices.clld.org/parameters}))}
}
\description{
This function downloads data from Vanuatu Voices (\url{https://vanuatuvoices.clld.org/}). You need the internet connection.
}
\seealso{

}
\author{

}
