% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vanuatu.feature.R
\name{vanuatu.feature}
\alias{vanuatu.feature}
\title{Download Vanuatu Voices data}
\usage{
vanuatu.feature(features, na.rm = TRUE)
}
\arguments{
\item{features}{A vector with parameters from Concepts (\url{https://vanuatuvoices.clld.org/parameters}))}

\item{na.rm}{Logical. If TRUE function removes all languages not available in lingtypology database. By default is TRUE.}
}
\description{
This function downloads data from Vanuatu Voices (\url{https://vanuatuvoices.clld.org/}). You need the internet connection.
}
\seealso{
\code{\link{abvd.feature}}, \code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{bivaltyp.feature}}, \code{\link{eurasianphonology.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{soundcomparisons.feature}}, \code{\link{uralex.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Mikhail Leonov
}
