% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\docType{data}
\name{wals}
\alias{wals}
\title{WALS's Language identifiers}
\format{
A data frame with 2950 rows and 2 variables:
\describe{
  \item{wals.code}{WALS language identifier}
  \item{glottocode}{Glottocode}
}
}
\usage{
wals
}
\description{
Language identifiers from WALS (\url{https://wals.info/}). This dataset is created for \code{\link{wals.feature}} function.
}
\keyword{datasets}
