% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs.feature.R
\name{wcs.feature}
\alias{wcs.feature}
\title{Opens data from the database of WCS Data Archives}
\usage{
wcs.feature(feature)
}
\arguments{
\item{feature}{Berlin and Kay for Database of Berlin and Kay; WCS for World Color Survey Database}
}
\description{
This function opens downloaded data from the database of WCS Data Archives (\url{https://www1.icsi.berkeley.edu/wcs/data.html}).
}
\examples{

wcs.feature()

}
\seealso{
\code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
