% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs.feature.R
\name{wcs.feature}
\alias{wcs.feature}
\title{Opens data from the database of WCS Data Archives}
\usage{
wcs.feature(feature)
}
\arguments{
\item{feature}{"Berlin and Kay" to show data collected by Berlin and Kay; "WCS" for World Color Survey Database. Both databases were created by asking native speakers of different languages to name each of 330 Munsell chips, shown in a constant, random order, and (2), exposed to a palette of these chips and asked to to pick out the best example(s) ("foci") of the major terms elicited in the naming task. Two databases are differ in number of languages as WCS Database is significantly larger.}
}
\description{
This function opens downloaded data from the database of WCS Data Archives (\url{https://www1.icsi.berkeley.edu/wcs/data.html}).
}
\examples{

wcs.feature("WCS")

}
\seealso{
\code{\link{afbo.feature}}, \code{\link{autotyp.feature}}, \code{\link{oto_mangueanIC.feature}}, \code{\link{phoible.feature}}, \code{\link{sails.feature}}, \code{\link{valpal.feature}}, \code{\link{wals.feature}}
}
\author{
Kirill Koncha <majortomblog@gmail.com>
}
