% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs.R
\docType{data}
\name{wcs}
\alias{wcs}
\title{WCS data}
\format{
A data frame with 819912 rows and 19 variables:
\describe{
  \item{id}{Language id}
  \item{speaker_id}{Speaker id}
  \item{term_abbr}{Term abbriviation}
  \item{language_original}{Language name}
  \item{language}{Another Language name}
  \item{family}{Language family}
  \item{iso}{ISO code}
  \item{glottocode}{Glottocode}
  \item{location}{Location}
  \item{fieldworkers}{Fieldworkers}
  \item{longitude}{Longitude}
  \item{latitude}{Latitude}
  \item{focus_response}{Focus response}
  \item{grid_coordinates}{Grid coordinates}
  \item{term_id}{Term id}
  \item{term_transcription}{Term transcription}
  \item{speakers_age}{Speakers age}
  \item{speakers_sex}{Speakers sex}
  \item{chip_range}{Chip range}
}
}
\usage{
wcs
}
\description{
Data from The database of WCS Data Archives (\url{https://www1.icsi.berkeley.edu/wcs/data.html}). This dataset is created for \code{\link{wcs.feature}} function.
}
\keyword{datasets}
