% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcs_bk.R
\docType{data}
\name{wcs_bk}
\alias{wcs_bk}
\title{WCS data}
\format{
A data frame with 32211 rows and 13 variables:
\describe{
  \item{language_id}{Language id}
  \item{speaker_id}{Speaker id}
  \item{term_abbr}{Color term abbriviation}
  \item{language_original}{Language name in the original dataset}
  \item{language}{Language name in lingtypology}
  \item{iso}{ISO code}
  \item{glottocode}{Glottocode}
  \item{longitude}{Longitude}
  \item{latitude}{Latitude}
  \item{focus_response}{Focus response}
  \item{grid_coordinates}{Grid coordinates}
  \item{term_id}{Color term id}
  \item{term_transcription}{Color term transcription}
}
}
\usage{
wcs_bk
}
\description{
Data from The database of WCS Data Archives (\url{https://www1.icsi.berkeley.edu/wcs/data.html}). This dataset is created for \code{\link{wcs.feature}} function.
}
\keyword{datasets}
